Added utility light control
Utility light (similar to Meade 497 utility light) is an LED and a series resistor  connected to one of the PWM outputs. The brightness can be controlled via SHC menu - does not get saved to NV memory for now. This was tested on ESP32 SHC, should work on Teensy too but will require pin mapping to suit.