Allow temperature updates when WEATHER is OFF
My JTW controller does not have a BME280 chip so WEATHER is set to OFF
as such firstScanDone never flips and the code that averages out temperature updates is never run.
this means that a :SX9A command does not update the temperature.

This allows for offline updates of weather parameters ( diffraction based tracking )